Move project sorting into ProjectPresentation class
