+++ package.json
@@ -1,6 +1,6 @@
 {
     "name": "sketchbook2",
-    "version": "0.0.5",
+    "version": "0.0.6",
     "private": true,
     "scripts": {
         "dev": "vite dev",
