--- src/lib/base/Util/ConfigTypes.ts
@@ -1,6 +1,0 @@
-// Possible sort orders for projects and groups (set defaults in config/config.ts)
-export enum SortOrder {
-    Alphabetical = 'alphabetical',
-    Chronological = 'chronological',
-    ReverseChronological = 'reverse-chronological'
-}
