+++ tests/component/GroupSelector.test.ts
@@ -2,7 +2,7 @@ import { render, fireEvent, screen, cleanup } from '@testing-library/svelte';
 import { describe, it, expect, afterEach } from 'vitest';
 import { type ProjectConfig, ProjectConfigDefaults } from '$lib/base/ProjectConfig/ProjectConfig';
 import GroupSelector from '$lib/components/GroupSelector.svelte';
-import { SortOrder } from '$lib/base/Util/ConfigTypes';
+import { SortOrder } from '$lib/base/FileLoading/ProjectPresentation';
 
 function projectWithGroups(groups: string[], date: number): ProjectConfig {
     const props = {} as ProjectConfig;
